refactor: reorganize seed and config files
- Delete app/core/config.py (dead re-export of settings)
- Rename app/core/seed.py → app/core/default_workspace.py to clarify its
  role (resolving the default workspace ID) vs. the seed script
- Move seed.py → scripts/seed.py for discoverability; fix internal output
  path accordingly
- Update Makefile and Dockerfile to reflect new script location
- Fix import sites in auth.py and routers/config.py
