--- backend/app/core/config.py
@@ -1,8 +1,0 @@
-"""
-Configuration module that provides access to application settings.
-This module re-exports the settings instance for use in the auth module.
-"""
-
-from app.settings import settings
-
-__all__ = ["settings"]
